Revert "Request metadata capture"
