+++ .gitignore
@@ -1,21 +1,6 @@
-# node & SvelteKit
 node_modules
-
-# Go 
 dist
-
-# dev data 
 mongo
-
-# git
 .gitcredentials
-
-# macOS 
 .DS_Store
-
-# ???
 .env
-
-# httpBin
-binManagerApp
-apiGateWayApp
