+++ api-gateway/adapters/bin-manager-grpc.go
@@ -1,0 +1,35 @@
+package adapters
+
+import (
+	"log"
+
+	"google.golang.org/grpc"
+	"google.golang.org/grpc/credentials/insecure"
+
+	"api-gateway/generated/adapters"
+)
+
+type BinMgmtConn struct {
+	Conn *grpc.ClientConn
+	Client *binManager.BinManagerClient
+}
+
+const binManagerSrvAddr = "bin-manager:65535"
+
+func ConnectToBinMgmtSrv() *BinMgmtConn {
+	conn, err := grpc.Dial(binManagerSrvAddr, grpc.WithTransportCredentials(insecure.NewCredentials()))
+	if err != nil {
+		log.Fatalf("did not connect to: %v", err)
+	}
+
+	client := binManager.NewBinManagerClient(conn)
+	return &BinMgmtConn{
+		Conn: conn,
+		Client: &client,
+	}
+}
+
+func (s *BinMgmtConn) CloseConn() {
+	s.Conn.Close()
+}
+
