+++ bin-manager/adapters/bin.proto
@@ -1,0 +1,39 @@
+/*
+Execute generate code from this protofile with the following command line:
+protoc --go_out=./generated --go_opt=paths=source_relative --go-grpc_out=./generated --go-grpc_opt=paths=source_relative adapters/bin.proto
+*/
+syntax = "proto3";
+package binManager;
+
+option go_package = "/binManager";
+
+message Params {}
+
+message NewBinResponse {
+  string binId = 1;
+}
+
+message LogRequestParams {
+  string binId = 1;
+  map<string, string> requestToLog = 2;
+}
+
+message LogRequestResponse {}
+
+message FetchBinContentsParams {
+  string binId = 1;
+}
+
+message HttpRequest {
+  map<string, string> contents = 1;
+}
+
+message FetchBinContentsResponse {
+  repeated HttpRequest binContents = 1;
+}
+
+service BinManager {
+  rpc GenerateNewBin(Params) returns (NewBinResponse) {}
+  rpc LogRequestToBin(LogRequestParams) returns (LogRequestResponse) {}
+  rpc FetchBinContents(FetchBinContentsParams) returns (FetchBinContentsResponse) {}
+};
