+++ bin-manager/server.go
@@ -5,9 +5,9 @@ import (
 	"log"
 	"net"
 
-	"bin-manager/internal/adapters"
-	db_service "bin-manager/internal/db-service"
-	binManagerGRPC "bin-manager/pkg/generated"
+	db_service "bin-manager/db-service"
+	"bin-manager/adapters"
+	"bin-manager/generated/adapters"
 
 	"google.golang.org/grpc"
 )
@@ -27,7 +27,7 @@ func main() {
 	grpcServer := grpc.NewServer()
 
 	srv := adapters.BinMgmtServer{}
-	binManagerGRPC.RegisterBinManagerServer(grpcServer, &srv)
+	binManager.RegisterBinManagerServer(grpcServer, &srv)
 
 	fmt.Printf("Starting Bin Management service on port %s\n", gpcPort)
 	if err := grpcServer.Serve(lis); err != nil {
