+++ httpBin.code-workspace
@@ -14,7 +14,7 @@
     }
   ],
   "settings": {
-    "workbench.colorTheme": "Atom One Dark",
+    "workbench.colorTheme": "Svelte Dark",
     "theme-by-language.useGlobalConfigOnly": false
   }
 }
